feat: Add pagination to popular movies and TV shows
This commit adds a new parameter to the GetPopularMovies and GetPopularTVShows methods, allowing for pagination of results. Additionally, all other methods that return slices of media now have a page parameter. A TODO has also been added to create a result struct wrapper that contains the results and total pages.