feat: Add pagination to media client API methods
This commit adds pagination to the following media client API methods:
- GetPopularMovies
- GetPopularTVShows
- GetRecentMovies
- SearchMovies
- SearchTVShows
- GetMoviesByGenre
- GetTVShowsByGenre

It also changes the return type of some methods from slices to structs containing a slice and additional information about the total number of pages and results. Finally, it updates the ToEpisodeRelease and ToMovieRelease methods to return pointers instead of values.