feat(BM-260): Replace Media's UUID by UINT on ID field
The Media model now has auto create and update time fields. The TmdbID field was removed from the Media model. The MediaID field in TvShow, Episode, and Movie models is now of type uint instead of string. The CategoryMedia and Rating models' primary key fields are now of type uint instead of string.