feat: Add pagination to GetTVShowsByActor method
This commit adds pagination functionality to the GetTVShowsByActor method in tmdb.go. The startIndex and endIndex of the TV shows are now calculated using math.Min() function, which ensures that the correct range is returned even if there are less than 20 TV shows left on the last page. The TotalPage field of PaginatedTVShowResults struct is now calculated using int(math.Round(float64(len(actorTVCredits.Cast)) / 20)), which rounds up the total number of pages.