feat: Add WatchListItems table to database
This commit adds a new table called "watch_list_item" to the database schema. The table has three columns: UserID, MediaID, and Status. Four new constants are also added to the MediaType enum: PLAN_TO_WATCH, WATCHING, FINISHED, and ABANDONED. Finally, the Migrate function in root.go is updated to include the new WatchListItems table in the list of tables that should be migrated when running migrations.