feat(transcoder): add intro video to main video
This commit adds an intro video to the main video by creating a temporary file to store the list of input files, writing the list of input files to the temporary file, and then concatenating them using ffmpeg. The main video is transcoded if it is not already encoded in h264. The function `ProcessFileTranscode` now takes an additional argument for the path of the intro video.