feat(transcoder): add intro file to audio extraction
This commit adds the ability to concatenate an intro file with the input audio streams during extraction. The `extractAudioStreams` function now takes an additional argument, `introFile`, which is used in the ffmpeg command as a filter_complex option. A log message has also been added to print out the full ffmpeg command being executed.