feat(transcoder): Add support for shifting subtitle timestamps
This commit adds the ability to shift subtitle timestamps by a specified duration. The `extractSubtitleStreams` function now takes an additional argument, `introFile`, which is used to obtain the duration of the intro video. The new functions `getVideoDuration` and `shiftSubtitleTimecodes` are added to get the duration of a video and shift subtitle timecodes respectively.