chore: update dependencies and log error message
This commit updates the indirect dependencies of the project. Specifically, it updates golang.org/x/net to version 0.0.0-20210610124326-52da8fb2a613 and golang.org/x/text to version 0.3.8. Additionally, this commit modifies transcoder.go to log an error message instead of returning an error when shifting subtitle timestamps fails in the extractSubtitleStreams function.