feat(cache): Add media cache implementation
This commit adds a new file `cache.go` which contains the implementation of a media cache. The media cache is used to store and retrieve movies, TV shows, episodes, search results, genres and actors. Each type of media has its own set of functions for adding and retrieving from the cache. The cache uses the `go-cache` package to manage expiration times for each item in the cache.