feat: Add Redis support to media cache
This commit adds Redis support to the media cache. A new implementation of the `mediaCache` interface has been added, called `redisMediaCache`. It uses a Redis client to store and retrieve data instead of an in-memory cache. The expiration time for each key is calculated based on the release date or air date of the corresponding movie, TV show, episode, or search result.