feat: Add Redis support to MediaClient
This commit adds Redis support to the MediaClient by creating a new constructor function `NewRedisMediaClient` that initializes a new instance of `mediaClient` with a Redis cache. The existing `cache` field has been changed from a pointer to an instance of `mediaCache`. A new implementation of the cache interface, `newRedisMediaCache`, has also been added.