feat: Add movie search by year
This commit adds the ability to search for movies by year in addition to the existing search functionality. The `SearchMoviesYear` function has been added to the `MediaClient` interface and implemented in both the `inMemoryMediaCache` and `redisMediaCache`. The necessary changes have also been made to the `tmdb.go` file, including adding a new method to the `MediaClient` interface and implementing it in the `mediaClient`.