refactor: Update default status for watch list items
The code changes update the default status for movie and TV show watch list items in the model.go file. The `Status` field is modified to have a default value of "PLAN_TO_WATCH" instead of being nullable. This change ensures that new watch list items are set to the default status automatically.