feat: Add actor search functionality
This commit adds the ability to search for actors matching a given query. The `SearchActors` method has been implemented in the `MediaClient` interface and the `mediaClient` struct. It takes a query string, page number, and an adult flag as parameters. The method makes use of the TMDB API to retrieve actor search results based on the provided query and options. The extracted results are then cached using the cache service.

The new `PaginatedActorResults` struct has been added to represent paginated actor search results. It contains fields for storing the total number of pages and total number of results, as well as a slice of `Actor` objects representing individual actors.

The `extractActors` function has also been added to extract relevant information from the TMDB API response and return a list of `Actor` objects.

Conventional commit rules have been followed for this commit message.